+++ src/pages/Pools/Home/Stats/MinCreateBond.tsx
