+++ src/pages/Pools/Home/Stats/MinJoinBond.tsx
